chore: use new ToolCode component and refactor JsonFormatter outside of components folder
